지갑 상세 조회 기능 추가

+++ lib/platform/apis/wallet.dart
@@ -23,7 +23,7 @@ class WalletApi {
     });
   }
 
-  static Future<Response> buyTik(double tikAmount) async {
+  static Future<Response> buyTik(int tikAmount) async {
     return await Api.client(serviceUrl: ServiceUrl.spendingWalletService).post(
       '/wallet/buy-tik',
       data: {
